requirements.txt
amqp==1.4.9
anyjson==0.3.3
apturl==0.5.2
asgiref==3.3.0
autopep8==1.4.4
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.3.0.23
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.2
django-admin==2.0.1
django-celery==3.3.1
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-faicon==0.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.12.0
django-phonenumber-field==5.2.0
django-six==1.0.4
django-smart-selects==1.5.9
django-versatileimagefield==2.0
djangorestframework==3.10.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kombu==3.0.37
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.0.1
platformdirs==2.0.2
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-magic==0.4.18
pytz==2019.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen==1.0.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
systemd-python==234
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.6.0
wadllib==1.3.3
xkit==0.0.0
xlwt==1.3.0
